Backup Before trying MultiThreading in Board, also added encapsulation improvements
SplitPane now has private static instances of board and sidePanel that
can be accessed from within the package by doing SidePanel.getBoard() or
SidePanel.getSidePanel to return the current instance of either one that
can then be used to access components they hold, such as the labels in
SidePanel.

Also, I'm going to try some big changes with multithreading the
collision updates and repaint updates inside of Board.java, so this
commit will be a revert point if I need.