Can add / remove entities now. GUI updates, stability fixes, minor changes
Lots of changes. Can add a default entitiy to the board (will extend
that later to be an icon-selection panel) and can delete entities, and
propertiesList will be updated accordingly. Only need to call
addEntity(x,y) or deleteEntitiy(index). Will probably overload these
methods but for now this is fine.
Also, changing an entity's name now updates its respective name in the
drop-down box.