Cannot set doubles equal to each other or 0
I hadn't noticed you were doing this; I learned in my java class last
semester that you should never compare doubles/floats to a single value
or 0 because it will never return true. This is because of precision
errors. Instead you're supposed to check if it's within a tiny range
(+/- .0001 for example)

I started fixing what I could see inside of Boundary but realized it's
all over the place, you're going to have to finish it.

There's a lot of (if doubleVar == 0) which will never return true
because 0 is an integer. Maybe if you did 0.0 but yeah precision error
so just do a range check.